+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Issue with loading related resources
 - Fix P2P sync progress showing greater than 100%
 - Background worker performance
+- The `requested` to `ready` task status update background worker 
 
 ### Changed
 - Refactored how the related resources SELECT and COUNT queries search results are represented.